// backend/apps/user-service/users/user.database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  users: {
    Tables: {
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          birth_date: string
          created_at: string | null
          dni: string
          email: string
          first_name: string
          id: string
          last_avatar_change: string | null
          last_password_change: string | null
          last_profile_update: string | null
          maternal_last_name: string
          paternal_last_name: string
          phone: string
          role: string
          updated_at: string | null
        }
        Insert: {
          avatar_url?: string | null
          birth_date: string
          created_at?: string | null
          dni: string
          email: string
          first_name: string
          id: string
          last_avatar_change?: string | null
          last_password_change?: string | null
          last_profile_update?: string | null
          maternal_last_name: string
          paternal_last_name: string
          phone: string
          role?: string
          updated_at?: string | null
        }
        Update: {
          avatar_url?: string | null
          birth_date?: string
          created_at?: string | null
          dni?: string
          email?: string
          first_name?: string
          id?: string
          last_avatar_change?: string | null
          last_password_change?: string | null
          last_profile_update?: string | null
          maternal_last_name?: string
          paternal_last_name?: string
          phone?: string
          role?: string
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}